+++ packages/terra-base/tests/wdio/base-spec.js
@@ -1,5 +1,7 @@
-/* global $, before, browser */
+/* global Terra, $, before, browser */
+
 describe('Base', () => {
+  before(() => browser.setViewportSize(Terra.viewports('tiny')[0]));
   describe('Switching Locales', () => {
     before(() => browser.url('/#/raw/tests/terra-base/base/switch-locale-base'));
     it('Displays a default en locale message', () => {
@@ -22,5 +24,16 @@ describe('Base', () => {
       browser.url('/#/raw/tests/terra-base/base/no-translations-loaded-string-base');
       expect($('div').getText()).to.not.contain('No Translations String');
     });
+  });
+
+  describe('Background decoration', () => {
+    beforeEach(() => browser.url('/#/raw/tests/terra-base/base/switch-locale-base'));
+
+    Terra.should.themeEachCustomProperty('#root',
+      {
+        '--terra-base-background-color': 'red',
+        '--terra-base-background-image': 'url("https://github.com/cerner/terra-core/raw/master/terra.png")',
+      },
+    );
   });
 });